requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv
passlib
python-jose
pydantic
python-multipart
fastapi-mail
PyJWT
pydantic[email]
slowapi
email-validator
sqlalchemy
pymysql
pydantic==2.11.10
fastapi-mail==1.5.0
requests
httpx
cryptography  # Required for MySQL 8.0 caching_sha2_password authentication
croniter
apscheduler
upstash-redis
fpdf2

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
email-validator




